Added CMake files and removed VS solution files

+++ src/color.cpp
@@ -18,10 +18,10 @@ Color::Color(Int16 r,Int16 g,Int16 b,Int16 a)
 }
 Color::Color(const std::string &str) {ustring::string_to_array<Int16,Int32>(str,&r,atoi,4);}
 Color::Color(const Vector3 &v)
-	: r(v.x *255.f),g(v.y *255.f),b(v.z *255.f)
+	: r(static_cast<Int16>(v.x *255.f)),g(static_cast<Int16>(v.y *255.f)),b(static_cast<Int16>(v.z *255.f))
 {}
 Color::Color(const Vector4 &v)
-	: r(v.x *255.f),g(v.y *255.f),b(v.z *255.f),a(v.w *255.f)
+	: r(static_cast<Int16>(v.x *255.f)),g(static_cast<Int16>(v.y *255.f)),b(static_cast<Int16>(v.z *255.f)),a(static_cast<Int16>(v.w *255.f))
 {}
 void Color::Set(const Color &col) {this->r = col.r; this->g = col.g; this->b = col.b; this->a = col.a;}
 Vector3 Color::ToVector3() const {return Vector3{r /255.f,g /255.f,b /255.f};}
